refactor: reorder Search component for lazyloading
